Move a bunch of running-related code into modal.runner
